Update File Structure
Update file structure to encapsulate API better and clean up where code
is located.
This will make refactoring/changing the way the file system operates
much easier
