+++ Examples/test/Makefile
@@ -5,12 +5,24 @@ CFLAGS = -g -Wall -Wno-format
 # C Sources
 C_SOURCES = \
 main.c \
-../../src/storfs.c
+../../src/core.c \
+../../src/crc.c \
+../../src/wear.c \
+../../src/mount.c \
+../../src/touch.c \
+../../src/fgets.c \
+../../src/fputs.c \
+../../src/fopen.c \
+../../src/mkdir.c \
+../../src/rewind.c \
+../../src/rm.c
+
 
 # C Includes
 C_INCLUDES = \
 -I../test \
--I../../include 
+-I../../include \
+-I../../src
 
 # Build Path
 BUILD_DIR = build
@@ -32,7 +44,7 @@ $(BUILD_DIR)/%.o: %.c  | $(BUILD_DIR)
 
 $(BUILD_DIR)/$(TARGET): $(C_OBJECTS)
 	$(CC) $(CFLAGS) $(C_INCLUDES) -o $(BUILD_DIR)/$(TARGET) $(C_OBJECTS)
-	
+
 $(BUILD_DIR):
 	mkdir $@
