+++ src/mkdir.c
@@ -1,0 +1,9 @@
+#include "core.h"
+
+storfs_err_t storfs_mkdir(storfs_t *storfsInst, char *pathToDir)
+{   
+    STORFS_LOGI(TAG, "Making Directory at %s", pathToDir);
+
+    return file_handling_helper(storfsInst, (storfs_name_t *)pathToDir, DIR_CREATE, NULL);
+}
+
